Remove appMain.cpp from frameReceiver
main() entry point is now in FrameReceiverApp to match FrameProcessor.
Updated CMake config and unit test config to match.
