+++ frameReceiver/src/FrameReceiverApp.cpp
@@ -5,6 +5,8 @@
  *      Author: Tim Nicholls, STFC Application Engineering Group
  */
 
+#include <signal.h>
+#include <iostream>
 #include <fstream>
 using namespace std;
 
@@ -338,3 +340,39 @@ void FrameReceiverApp::stop(void)
   controller_->stop();
 }
 
+//! Interrupt signal handler
+
+void intHandler (int sig)
+{
+  FrameReceiver::FrameReceiverApp::stop ();
+}
+
+//! Main application entry point
+
+int main (int argc, char** argv)
+{
+  int rc = 0;
+
+  // Trap Ctrl-C and pass to interrupt handler
+  signal (SIGINT, intHandler);
+  signal (SIGTERM, intHandler);
+
+  // Set the application path and locale for logging
+  setlocale(LC_CTYPE, "UTF-8");
+  OdinData::app_path = argv[0];
+
+  // Create a FrameReceiverApp instance
+  FrameReceiver::FrameReceiverApp fr_instance;
+
+  // Parse command line arguments and set up node configuration
+  rc = fr_instance.parse_arguments (argc, argv);
+
+  if (rc == 0)
+  {
+    // Run the instance
+    fr_instance.run ();
+  }
+
+  return rc;
+
+}
