+++ frameReceiver/test/FrameReceiverUnitTestMain.cpp
@@ -5,6 +5,9 @@
  *      Author: tcn45
  */
 
+#include "DebugLevelLogger.h"
+IMPLEMENT_DEBUG_LEVEL;
+
 #define BOOST_TEST_MODULE "FrameReceiverUnitTests"
 #define BOOST_TEST_MAIN
 #include <boost/test/unit_test.hpp>
